requirements.txt
fire
pytorch
torchvision
